+++ strict-mutable-base/CHANGELOG.md
@@ -1,2 +1,5 @@
+# strict-mutable-base-1.1.0.0 (????-??-??)
+* Rename `getChanContents'` to `getChan'Contents` for consistency.
+
 # strict-mutable-base-1.0.0.0 (2024-09-02)
 * Initial release.